SDL3: Rebuild all bridged pages from scratch with wikiheaders.
This cleaned out a bunch of cruft in the files and got rid of now-bogus pages
for renamed/removed APIs.

The changes were then manually sifted through, to keep (most) code examples
and a few other details.

Custom categories are lost for now, but there is pending work to have
wikiheaders automatically add categories based on header name.